feat: add local profile to simplify the local run of the app
