+++ src/main/resources/application-local.properties
@@ -1,0 +1,8 @@
+spring.data.mongodb.uri=mongodb://localhost:27017/
+spring.data.mongodb.database=pancake-nft-publish
+spring.task.scheduling.pool.size=3
+
+nft.bunny.collection.last.index=960740
+
+wallet.address=PUT_HERE_SOME-TEST-WALLET
+secure.token=PUT_HERE_SOME-TEST-SECURE_TOKEN
